requirements.txt
certifi==2019.6.16
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.2.4
django-appconf==1.0.3
django-bootstrap4==0.0.8
django-common-helpers==0.9.2
django-cors-headers==3.1.0
django-cron==0.5.1
django-ipware==2.1.0
django-modeltranslation==0.13.3
django-rest-swagger==2.2.0
djangorestframework==3.10.2
djangorestframework-recaptcha==0.2.0
idna==2.8
itypes==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
openapi-codec==1.3.2
Pillow==6.1.0
psycopg2==2.8.3
pytz==2019.2
requests==2.22.0
simplejson==3.16.0
six==1.12.0
sqlparse==0.3.0
uritemplate==3.0.0
urllib3==1.25.3
